% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moter.R
\docType{package}
\name{moter}
\alias{moter}
\title{moter: A package for creating morphological tessellations from building
footprint polygons}
\description{
The moter package is designed to explore the morphological tessellation
algorithm described by Fleischmann et al. (2020) and implemented in the
Python package Momepy \url{https://docs.momepy.org}.
}
\details{
Morphological tessellations are an approach to partitioning space based on
voronoi polygons created around 2D building footprints. These resulting
morphological tessellation polygons approximate a plot scale area and can be
used in quantitative study of urban form and morphology.
}
