% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{motess}
\alias{motess}
\title{Create morphological tessellations}
\usage{
motess(X, unique_id, limit, shrink = 0.4, segment = 0.5, verbose = TRUE)
}
\arguments{
\item{X}{Spatial polygons of building footprints.}

\item{unique_id}{String of a column name in \code{X} with unique IDs.}

\item{limit}{Spatial polygon defining the study area.}

\item{shrink}{Distance for inward buffer on building footprints. Default 0.4.}

\item{segment}{Distance (in meters) between points created on building
footprint segments. Default 0.5.}

\item{verbose}{Display progress. Default True.}
}
\value{
A spatial dataset in \code{sf} format of the tessellation polygons.
}
\description{
Create morphological tessellations
}
\details{
The algorithm is described by Fleischmann et al. (2020).
}
\examples{

}
\references{
Fleischmann, M., A. Feliciotti, O. Romice, S. Porta (2020).
"Morphological tessellation as a way of partitioning space: Improving
consistency in urban morphology at the plot scale." \emph{Computers,
Environment, & Urban Systems}, 80,
101441.\url{https://doi.org/10.1016/j.compenvurbsys.2019.101441}
}
\seealso{
\url{https://docs.momepy.org}
}
